wm: changes to cwm in order to allow xmobar
cwm has no status bar like i3 does with i3status, but it's possible to
use any general purpose status bar like xmobar. A gap in top side was
added in order to open a space to place xmobar and the xmobarrc was
added as well. It surely needs some tweaks, but it's enough for now.