mutt: changed some config vars default value
The most important change here is the date_format and index_format vars:
$date_format is used in $attribute var that is the text used before
quoted emails in replies, hence the date format being used was not
following the standard that includes, day, month, year and timezone.
Removing $date_format (setting it back to the default value) and setting
the $index_format date format manually solved the problem.