email: use mbsync instead of offlineimap
Offlineimap was consuming too much from the processor for simple syncs, because
of that mbsync is being used as default instead. It consumes far less system
resources for both create new maildirs and sync with the existing ones.

With that, some changes were made on muttrc in order to accomodate some of the
webui filters. Also small corrections were made regarding files pathing.