mutt: fix editor= option when using xterm
Insteresting enough when using xterm -e option the $EDITOR option is empty
while it should have the editor's name, as it's set in the normal environment.
Not sure if something changed in xterm and the env vars are now being erased.
But the thing is that in mutt the 'ed' editor was being used instead of the one
stated in $EDITOR and because of that now I'm setting the editor= option
directly.