nvim/lazy: limit concurrent processes
When using unlimited number of concurrent tasks to update/fetch/... with
Lazy I was facing some issues during git-fetch authentication, which might
be a problem/limitation on the gpg-agent trying to deal with too many
requests for SSH key authentication or with git-fetch itself. Regardless the
culprit, limiting it to a number like 8 got rid of the issue.